Created journal component using the code @eyanka
