+++ src/pages/PromptJournal.js
@@ -1,20 +1,8 @@
 import React from 'react';
-import { Link } from 'react-router-dom';
-import GeneratePrompt from '../js/Prompt.js';
-import Button from '../js/Button.js';
+import JournalContainer from '../js/JournalContainer';
 
 const PromptJournal = () => {
-  const fillerFunc = () => {
-    console.log('button test');
-  };
-
-  return (
-    <div className="journal-entry">
-      <GeneratePrompt />
-      <Link to="/">go back</Link>
-      <Button type="button" name="save" text="Save" class="btn btn-primary" onclick={fillerFunc} />
-    </div>
-  );
+  return <JournalContainer />;
 };
 
 export default PromptJournal;
