+++ src/styles/main.css
@@ -564,6 +564,15 @@ body{
   --tw-text-opacity: 1;
   color: rgb(0 44 65 / var(--tw-text-opacity));
 }
+.btn-container{
+  display: flex;
+  justify-content: flex-end;
+}
+.btn-container > :not([hidden]) ~ :not([hidden]){
+  --tw-space-x-reverse: 0;
+  margin-right: calc(0.625rem * var(--tw-space-x-reverse));
+  margin-left: calc(0.625rem * calc(1 - var(--tw-space-x-reverse)));
+}
 .shuffle-icon{
   height: 1.75rem;
   width: 1.75rem;
@@ -571,22 +580,69 @@ body{
   --tw-text-opacity: 1;
   color: rgb(14 137 146 / var(--tw-text-opacity));
 }
-.journal-entry{
+.journal-container{
   height: 100vh;
+  width: 100vw;
+}
+.journal-container > :not([hidden]) ~ :not([hidden]){
+  --tw-space-y-reverse: 0;
+  margin-top: calc(0.625rem * calc(1 - var(--tw-space-y-reverse)));
+  margin-bottom: calc(0.625rem * var(--tw-space-y-reverse));
+}
+.journal-container{
   background-color: rgb(255 255 255 / var(--tw-bg-opacity));
-  --tw-bg-opacity: 0.5;
-  padding: 1rem;
+  --tw-bg-opacity: 0.25;
+}
+.journal-entry{
+  margin-left: auto;
+  margin-right: auto;
+  width: 80%;
+  padding: 4rem;
+}
+.text-entry > :not([hidden]) ~ :not([hidden]){
+  --tw-space-y-reverse: 0;
+  margin-top: calc(0.625rem * calc(1 - var(--tw-space-y-reverse)));
+  margin-bottom: calc(0.625rem * var(--tw-space-y-reverse));
+}
+.text-entry{
+  border-radius: 1.5rem;
+  border-width: 1px;
+  border-style: solid;
+  --tw-border-opacity: 1;
+  border-color: rgb(0 44 65 / var(--tw-border-opacity));
+  --tw-bg-opacity: 1;
+  background-color: rgb(246 239 222 / var(--tw-bg-opacity));
+  padding: 2.5rem;
+  --tw-shadow: 0 4px 6px -1px rgb(0 0 0 / 0.1), 0 2px 4px -2px rgb(0 0 0 / 0.1);
+  --tw-shadow-colored: 0 4px 6px -1px var(--tw-shadow-color), 0 2px 4px -2px var(--tw-shadow-color);
+  box-shadow: var(--tw-ring-offset-shadow, 0 0 #0000), var(--tw-ring-shadow, 0 0 #0000), var(--tw-shadow);
 }
 .prompt{
-  font-size: 2.25rem;
-  line-height: 2.5rem;
+  font-size: 1.875rem;
+  line-height: 2.25rem;
+  font-weight: 600;
+}
+.prompt-container{
+  display: flex;
+  flex-direction: row;
+}
+.prompt-container > :not([hidden]) ~ :not([hidden]){
+  --tw-space-x-reverse: 0;
+  margin-right: calc(0.5rem * var(--tw-space-x-reverse));
+  margin-left: calc(0.5rem * calc(1 - var(--tw-space-x-reverse)));
 }
 .subheading{
   font-size: 1.25rem;
   line-height: 1.75rem;
 }
-.block{
+.title-input{
   display: block;
+  width: 100%;
+  --tw-bg-opacity: 1;
+  background-color: rgb(246 239 222 / var(--tw-bg-opacity));
+  padding: 0.25rem;
+  font-size: 1.5rem;
+  line-height: 2rem;
 }
 .flex{
   display: flex;
@@ -618,4 +674,14 @@ body {
   background-repeat: no-repeat;
   background-size: 100% auto;
   background-position: bottom;
+  background-attachment: fixed;
+}
+
+.ql-container.ql-snow {
+  height: 25vh;
+  max-height: 25vh;
+}
+
+.ql-editor {
+  overflow: auto;
 }
