+++ src/styles/tailwind.css
@@ -14,6 +14,16 @@ body {
   background-repeat: no-repeat;
   background-size: 100% auto;
   background-position: bottom;
+  background-attachment: fixed;
+}
+
+.ql-container.ql-snow {
+  height: 25vh;
+  max-height: 25vh;
+}
+
+.ql-editor {
+  overflow: auto;
 }
 
 @layer base {
@@ -35,19 +45,39 @@ body {
     @apply bg-white text-custom-teal-dark border border-solid border-custom-teal-dark;
   }
 
+  .btn-container {
+    @apply flex justify-end space-x-2.5;
+  }
+
   .shuffle-icon {
     @apply text-custom-teal-base bg-transparent h-7 w-7;
   }
 
+  .journal-container {
+    @apply bg-white bg-opacity-25 w-screen h-screen space-y-2.5;
+  }
+
   .journal-entry {
-    @apply bg-white bg-opacity-50 h-screen p-4;
+    @apply w-4/5 p-16 mx-auto;
+  }
+
+  .text-entry {
+    @apply bg-custom-cream rounded-3xl border border-solid border-custom-teal-dark shadow-md p-10 space-y-2.5;
   }
 
   .prompt {
-    @apply text-4xl;
+    @apply text-3xl font-semibold;
+  }
+
+  .prompt-container {
+    @apply flex flex-row space-x-2;
   }
 
   .subheading {
     @apply text-xl;
+  }
+
+  .title-input {
+    @apply bg-custom-cream block text-2xl w-full p-1;
   }
 }
